refactor: 월수입 페이지, 입력 컴포넌트와 병합
